requirements.txt
flask
mysql-connector
jquery
jsonify